requirements.txt
certifi==2018.11.29
chardet==3.0.4
idna==2.8
Pillow==5.4.1
requests==2.21.0
urllib3==1.24.1
filetype==1.0.5
